requirements.txt
lxml==4.9.3
numpy==1.26.2
pandas==2.1.3
PySide6==6.6.0
PySide6-Addons==6.6.0
PySide6-Essentials==6.6.0
python-dateutil==2.8.2
python-docx==1.1.0
pytz==2023.3.post1
shiboken6==6.6.0
six==1.16.0
typing_extensions==4.8.0
tzdata==2023.3
